+++ cores/epoxy/pgmspace.h
@@ -36,6 +36,7 @@
 #define strcasecmp_P strcasecmp
 #define strchr_P strchr
 #define strrchr_P strrchr
+#define strstr_P strstr
 #define memcpy_P memcpy
 #define snprintf_P snprintf
 #define vsnprintf_P vsnprintf