+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 * Unreleased
+* 1.0 (2021-09-30)
+    * Add `epoxy_argc` and `epoxy_argv` as extern global variables which
+      are set to the `argc` and `argv` parameters passed into the global
+      `main()`.
+        * Allows command line arguments to be passed into an Arduino
+          application, to facilitate debugging on a Unix desktop machine.
+        * Add `examples/CommandLine` which provides a basic command line parser
+          that can be copied and modified for other applications.
+    * Add `toString()` to `IPAddress` class, activated with `EPOXY_CORE_ESP8266`
+      for compatibility with ESP8266 Core.
+    * Add `strstr_P()` to `pgmspace.h`.
+    * Finally fix [Issue #2](https://github.com/bxparks/EpoxyDuino/issues/2) and
+      [Issue #25](https://github.com/bxparks/EpoxyDuino/issues/25).
 * 0.8 (2021-08-08)
     * Add `EpoxyMockTimerOne` mock library for `TimerOne`
       (https://github.com/PaulStoffregen/TimerOne).